+++ conf/script/src/build_system/build_target/__init__.py
@@ -1,2 +1,2 @@
-from .build_target import *
+from .build_target_cls import *
 from .compiler_instance_targets import *
